feat: add UnreadCountCalculator

+++ src/main/kotlin/com/github/cloverchatserver/domain/chatmsg/business/data/ChatMessageCreation.kt
@@ -10,7 +10,7 @@ data class ChatMessageCreation(
     val createUserId: Long,
     val content: String
 ) {
-    fun toChatMessage(chatRoom: ChatRoom, createdBy: Account, num: Long): ChatMessage {
+    fun toChatMessage(chatRoom: ChatRoom, createdBy: Account, num: Int): ChatMessage {
         return ChatMessage(null, chatRoom, createdBy, content, LocalDateTime.now(), num)
     }
 }
